Add step to ensure removed files are cleared.
Hugo does not compare the destination with the source, so it doesn't
know when existing files in the destination need to be removed.
The --cleanDestinationDir flag only concerns static files, so it
doesn't fix this problem.

Instead, the destination is first cleared of everything except hidden
files (that is, .git). This gives Hugo a clean slate to build on.